APIs/openchannel.io/market/2.0.24/python-aiohttp-client/tests/test_files_upload_files_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web
from aiohttp import FormData

from openapi_server.models.file import File
from openapi_server.models.file_download import FileDownload


pytestmark = pytest.mark.asyncio

async def test_files_by_id_or_url_get(client):
    """Test case for files_by_id_or_url_get

    Get the details for a file.
    """
    params = [('fileIdOrUrl', 'file_id_or_url_example')]
    headers = { 
        'Accept': '*/*',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v2/files/byIdOrUrl',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_files_download_get(client):
    """Test case for files_download_get

    A signed URL for downloading a private file can be returned by providing the fileId.
    """
    params = [('fileId', 'file_id_example'),
                    ('validSeconds', 56)]
    headers = { 
        'Accept': '*/*',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v2/files/download',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_files_get(client):
    """Test case for files_get

    Returns a paginated list of files
    """
    params = [('query', 'query_example'),
                    ('sort', 'sort_example'),
                    ('pageNumber', 56),
                    ('limit', 56)]
    headers = { 
        'Accept': '*/*',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v2/files',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("multipart/form-data not supported by Connexion")
async def test_files_post(client):
    """Test case for files_post

    Uploads a file.
    """
    params = [('isPrivate', True),
                    ('hash', 'hash_example')]
    headers = { 
        'Accept': '*/*',
        'Content-Type': 'multipart/form-data',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = FormData()
    data.add_field('file', '/path/to/file')
    response = await client.request(
        method='POST',
        path='/v2/files',
        headers=headers,
        data=data,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_files_url_post(client):
    """Test case for files_url_post

    Uploads a file from a URL
    """
    params = [('url', 'url_example'),
                    ('isPrivate', True)]
    headers = { 
        'Accept': '*/*',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='POST',
        path='/v2/files/url',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

